Move random string generation out of testing
It's used in more than testing, so put it somewhere where the app and
testing infra can access it.
