+++ app/models/user.py
@@ -1,6 +1,6 @@
 import time
 
-import util.testing
+import util.string
 from modern_paste import db
 
 
@@ -32,7 +32,7 @@ class User(db.Model):
         self.password_hash = password_hash
         self.name = name
         self.email = email
-        self.api_key = util.testing.random_alphanumeric_string(length=64)
+        self.api_key = util.string.random_alphanumeric_string(length=64)
 
     @staticmethod
     def is_authenticated():
